fix: silence MSVC C4310 (cast truncates constant) in VgmExporter
Windows CI builds with /WX (warnings as errors). Casting 0xB4/0xB3
(>127) directly to char triggers C4310 on MSVC because char is signed;
clang doesn't warn on the equivalent cast, so this only surfaced on
Windows CI. Route through uint8_t first, matching how every other byte
constant in this file already reaches char/uint8_t.

+++ Core/Shared/Audio/VgmExporter.cpp
@@ -176,13 +176,13 @@ void VgmExporter::WriteChipCommand(VgmChip chip, uint8_t addrOrPort, uint8_t val
 {
 	switch(chip) {
 		case VgmChip::NesApu:
-			_stream.put((char)0xB4);
+			_stream.put((char)(uint8_t)0xB4);
 			_stream.put((char)addrOrPort);
 			_stream.put((char)value);
 			break;
 
 		case VgmChip::GameBoyDmg:
-			_stream.put((char)0xB3);
+			_stream.put((char)(uint8_t)0xB3);
 			_stream.put((char)addrOrPort);
 			_stream.put((char)value);
 			break;
